requirements.txt
streamlit
pandas
gspread
google-auth
openpyxl
xlsxwriter